refactor: 숫자 생성기 service 패키지로 이동
