+++ src/main/java/racingcar/service/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.service;
 
 public interface NumberGenerator {
