+++ src/main/java/racingcar/service/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
