+++ src/test/java/racingcar/service/RandomNumberAcceleratorTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.car.Car;
-import racingcar.util.MockNumberGenerator;
 
 class RandomNumberAcceleratorTest {
